requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.7.0
av==14.0.1
babel==2.16.0
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.0.1
bs4==0.0.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.1
ctranslate2==4.5.0
cycler==0.12.1
cymem==2.0.10
dataclasses-json==0.6.7
dateparser==1.1.8
diskcache==5.6.3
docopt==0.6.2
editdistance==0.8.1
elevenlabs==1.13.3
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
epitran==1.17
espeak_phonemizer==1.3.1
faster-whisper==1.1.0
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.10.0
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_en==2.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
huggingface-hub==0.26.5
humanfriendly==10.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonlines==1.2.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.26.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown2==2.5.2
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
munkres==1.1.4
murmurhash==1.0.11
mypy-extensions==1.0.0
networkx==3.4.2
num2words==0.5.13
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.20.1
packaging==24.2
panphon==0.21.2
phonemes2ids==1.2.2
pillow==11.0.0
preshed==3.0.9
priority==2.0.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
pytz==2024.2
PyYAML==6.0.2
Quart==0.19.9
quart-cors==0.7.0
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.11.0
srsly==2.4.8
swagger-ui-py==21.12.8
sympy==1.13.1
thinc==8.3.2
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
xdgenvpy==2.4.0
